:star: Inflate database forms on main activity and upload them to the service after click refresh button

+++ app/src/main/java/com/treecutforms/models/DatabaseForm.java
@@ -15,10 +15,11 @@ public class DatabaseForm extends RealmObject {
     private Form form;
     private boolean uploaded;
 
+
     public DatabaseForm() {
     }
 
-    public DatabaseForm(int id, Form form, boolean uploaded) {
+    public DatabaseForm(long id, Form form, boolean uploaded) {
         this.id = id;
         this.form = form;
         this.uploaded = uploaded;
